requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
django-crispy-forms==1.7.2
djangorestframework==3.9.0
idna==2.7
Pillow==5.3.0
pytz==2018.6
requests==2.20.0
urllib3==1.24.1
